Changed the efficient bridge-finding algorithm
Got rid of the unnecessary visited and parent array to make it more
space efficient. I also used the idea that the low array is the lowest
numbered vertex that can be reached, so a depth first search can
initiate at a connected component by passing the vertex itself in as its
parent.